Check if a stack exists before delete

+++ index.js
@@ -230,12 +230,20 @@ config.deleteStack = function(options, callback) {
         region: options.region
     }));
 
-    confirmAction('Ready to delete the stack ' + options.name + '?', function (confirm) {
-        if (!confirm) return callback();
-        cfn.deleteStack({
-            StackName: options.name
-        }, callback);
-    })
+    cfn.describeStacks({ StackName: options.name }, function (err, data) {
+        if (err) return callback(err);
+        var status = data.Stacks[0].StackStatus;
+        if (status === 'DELETE_COMPLETE' || status === 'DELETE_IN_PROGRESS') {
+            return callback(new Error([options.name, status].join(' ')));
+        }
+        
+        confirmAction('Ready to delete the stack ' + options.name + '?', function (confirm) {
+            if (!confirm) return callback();
+            cfn.deleteStack({
+                StackName: options.name
+            }, callback);
+        });
+    });
 };
 
 config.stackInfo = function(options, callback) {
